+++ src/ymm/cli.py
@@ -14,25 +14,13 @@ parser.add_argument('-f','--file', default=DEFAULT_FILE,
                     help='YAML file of actions')
 parser.add_argument('-d','--debug', action='store_true',
                     help='print debugging information')
+parser.add_argument('-l','--list', action='store_true',
+                    help='list available actions')
 parser.add_argument('-v','--version', action='version',
                     version=f'%(prog)s {__version__}')
 
-def add_versions(ctx):
-    try:
-        local_version = Path(VERSION_FILE).read_text().strip()
-    except e:
-        print(e)
-        return ctx
-    digits = re.findall(r'\d+', local_version)
-    last = int(digits[-1])
-    devNext = local_version.replace(f'dev{last}',f'dev{last+1}')
-    ctx['__version__'] = local_version
-    ctx['__version_digits__'] = digits
-    ctx['__version_next__'] = devNext
-    return ctx
-
 def context(args):
-    ctx = add_versions(dict(os.environ))
+    ctx = dict(os.environ)
     for arg in vars(args):
         value = getattr(args, arg)
         ctx[arg] = getattr(args, arg)
@@ -42,10 +30,13 @@ def main():
     args = parser.parse_args()
     file = args.file
     ymm = load_file(file)
+    if args.list:
+        print('actions:')
+        for action in ymm.actions(): print(f'    - {action}')
     ymm.env = context(args)
     actions = args.actions
     for action in actions:
-        ymm.log(f'; {action}')
+        print(f'; {action}')
         results = ymm.run(action)
 
 #main()