+++ src/ymm/ymm.py
@@ -6,23 +6,38 @@ class YMM:
     def __init__(self, yaml, debug=True):
         self.yaml = yaml
         self.env = {}
+        self.i = 0
 
-    def run(self,arg=False):
-        if not arg in self.yaml: sys.exit('ERROR: action [{arg}] not found')
-        actions = self.yaml[arg]
-        results = [self.execute(cmd) for cmd in actions]
+    def actions(self):
+        return list(self.yaml.keys())
+
+    def run(self,action=False):
+        if not action in self.yaml: sys.exit(f'ERROR: action [{action}] not found')
+        commands = self.yaml[action]
+        self.action = action
+        self.i = 0
+        results = [self.execute(cmd) for cmd in commands]
         return results
 
     def execute(self, cmd):
         if cmd[0] == kCall: return "\n".join(self.run(cmd[1:]))
         sub = cmd.format(**self.env)
-        args = sub.split(" ")
-        self.log(args)
-        result = subprocess.run(args, stdout=subprocess.PIPE)
+        commands = sub.split(" ")
+        self.log(commands)
+        result = subprocess.run(commands, stdout=subprocess.PIPE)
         msg = result.stdout.decode("utf-8").strip()
-        print(f'# {msg}')
-        self.env[kLast]=msg
+        self.save(msg)
         return msg
 
-    def log(self, arg):
-        if self.env.get(kLog, False): print(f'YMM.log {arg}')
+    def save(self, msg):
+        self.i += 1
+        key = f'{self.action}#{self.i}'
+        print(f'  {key}: {msg}')
+
+        if kLast in self.env: self.env[kPrior] = self.env[kLast]
+        self.env[kLast]=msg
+        self.env[key]=msg
+        return msg
+
+    def log(self, action):
+        if self.env.get(kLog, False): print(f'YMM.log {action}')