+++ CHANGELOG.md
@@ -1,7 +1,17 @@
-0.4.0 (unreleased)
+0.4.1 (unreleased)
 ------------------
 
 - Nothing changed yet.
+
+
+0.4.0 (2022-03-07)
+------------------
+
+- kPrior
+- undo version hack
+- always echo action
+- store results
+- keyed output
 
 
 0.3.0 (2022-03-05)